+++ commonroad/scenario/lanelet.py
@@ -88,6 +88,38 @@ class StopLine:
         self._traffic_sign_ref = traffic_sign_ref
         self._traffic_light_ref = traffic_light_ref
 
+    def __eq__(self, other):
+        if not isinstance(other, StopLine):
+            warnings.warn(f"Inequality between StopLine {repr(self)} and different type {type(other)}")
+            return False
+
+        prec = 10
+        start_string = np.array2string(np.around(self._start.astype(float), prec), precision=prec)
+        start_other_string = np.array2string(np.around(other.start.astype(float), prec), precision=prec)
+        end_string = np.array2string(np.around(self._end.astype(float), prec), precision=prec)
+        end_other_string = np.array2string(np.around(other.end.astype(float), prec), precision=prec)
+
+        if start_string == start_other_string and end_string == end_other_string \
+                and self._line_marking == other.line_marking and self._traffic_sign_ref == other.traffic_sign_ref \
+                and self._traffic_light_ref == other.traffic_light_ref:
+            return True
+
+        warnings.warn(f"Inequality of StopLine {repr(self)} and the other one {repr(other)}")
+        return False
+
+    def __hash__(self):
+        start_string = np.array2string(np.around(self._start.astype(float), 10), precision=10)
+        end_string = np.array2string(np.around(self._end.astype(float), 10), precision=10)
+        return hash((start_string, end_string, self._line_marking, frozenset(self._traffic_sign_ref),
+                     frozenset(self._traffic_light_ref)))
+
+    def __str__(self):
+        return f'StopLine from {self._start} to {self._end}'
+
+    def __repr__(self):
+        return f"StopLine(start={self._start.tolist()}, end={self._end.tolist()}, line_marking={self._line_marking}, " \
+               f"traffic_sign_ref={self._traffic_sign_ref}, traffic_light_ref={self._traffic_light_ref})"
+
     @property
     def start(self) -> np.ndarray:
         return self._start
@@ -148,9 +180,6 @@ class StopLine:
         tmp = t_m.dot(np.vstack((line_vertices.transpose(), np.ones((1, line_vertices.shape[0])))))
         tmp = tmp[0:2, :].transpose()
         self._start, self._end = tmp[0], tmp[1]
-
-    def __str__(self):
-        return f'StopLine from {self._start} to  {self._end}'
 
 
 class Lanelet:
@@ -278,6 +307,74 @@ class Lanelet:
             self._traffic_lights = set()
         else:
             self.traffic_lights = traffic_lights
+
+    def __eq__(self, other):
+        if not isinstance(other, Lanelet):
+            warnings.warn(f"Inequality between Lanelet {repr(self)} and different type {type(other)}")
+            return False
+
+        list_elements_eq = self._stop_line == other.stop_line
+
+        lanelet_eq = True
+        polylines = [self._left_vertices, self._right_vertices, self._center_vertices]
+        polylines_other = [other.left_vertices, other.right_vertices, other.center_vertices]
+
+        for i in range(0, len(polylines)):
+            polyline = polylines[i]
+            polyline_other = polylines_other[i]
+            polyline_string = np.array2string(np.around(polyline.astype(float), 10), precision=10)
+            polyline_other_string = np.array2string(np.around(polyline_other.astype(float), 10), precision=10)
+            lanelet_eq = lanelet_eq and polyline_string == polyline_other_string
+
+        if lanelet_eq and self.lanelet_id == other.lanelet_id \
+                and self._line_marking_left_vertices == other.line_marking_left_vertices \
+                and self._line_marking_right_vertices == other.line_marking_right_vertices \
+                and set(self._predecessor) == set(other.predecessor) and set(self._successor) == set(other.successor) \
+                and self._adj_left == other.adj_left and self._adj_right == other.adj_right \
+                and self._adj_left_same_direction == other.adj_left_same_direction \
+                and self._adj_right_same_direction == other.adj_right_same_direction \
+                and self._lanelet_type == other.lanelet_type and self._user_one_way == self.user_one_way \
+                and self._user_bidirectional == other.user_bidirectional \
+                and self._traffic_signs == other.traffic_signs and self._traffic_lights == other.traffic_lights:
+            return list_elements_eq
+
+        warnings.warn(f"Inequality of Lanelet {repr(self)} and the other one {repr(other)}")
+        return False
+
+    def __hash__(self):
+        polylines = [self._left_vertices, self._right_vertices, self._center_vertices]
+        polyline_strings = []
+        for polyline in polylines:
+            polyline_string = np.array2string(np.around(polyline.astype(float), 10), precision=10)
+            polyline_strings.append(polyline_string)
+
+        elements = [self._predecessor, self._successor, self._lanelet_type, self._user_one_way,
+                    self._user_bidirectional, self._traffic_signs, self._traffic_lights]
+        frozen_elements = [frozenset(e) for e in elements]
+
+        return hash((self._lanelet_id, tuple(polyline_strings), self._line_marking_left_vertices,
+                     self._line_marking_right_vertices, self._stop_line, self._adj_left, self._adj_right,
+                     self._adj_left_same_direction, self._adj_right_same_direction, tuple(frozen_elements)))
+
+    def __str__(self):
+        return f"Lanelet with id {self._lanelet_id} has predecessors {set(self._predecessor)}, successors " \
+               f"{set(self._successor)}, left adjacency {self._adj_left} with " \
+               f"{'same' if self._adj_left_same_direction else 'opposite'} direction, and " \
+               f"right adjacency with {'same' if self._adj_right_same_direction else 'opposite'} direction"
+
+    def __repr__(self):
+        return f"Lanelet(left_vertices={self._left_vertices.tolist()}, " \
+               f"center_vertices={self._center_vertices.tolist()}, " \
+               f"right_vertices={self._right_vertices.tolist()}, lanelet_id={self._lanelet_id}, " \
+               f"predecessor={self._predecessor}, successor={self._successor}, adjacent_left={self._adj_left}, " \
+               f"adjacent_left_same_direction={self._adj_left_same_direction}, adjacent_right={self._adj_right}, " \
+               f"adjacent_right_same_direction={self._adj_right_same_direction}, " \
+               f"line_marking_left_vertices={self._line_marking_left_vertices}, " \
+               f"line_marking_right_vertices={self._line_marking_right_vertices}), " \
+               f"stop_line={repr(self._stop_line)}, lanelet_type={self._lanelet_type}, " \
+               f"user_one_way={self._user_one_way}, " \
+               f"user_bidirectional={self._user_bidirectional}, traffic_signs={self._traffic_signs}, " \
+               f"traffic_lights={self._traffic_lights}"
 
     @property
     def distance(self) -> np.ndarray:
@@ -958,9 +1055,6 @@ class Lanelet:
         else:
             return set()
 
-    def __str__(self):
-        return 'Lanelet with id:' + str(self.lanelet_id)
-
 
 class LaneletNetwork(IDrawable):
     """
@@ -1005,6 +1099,44 @@ class LaneletNetwork(IDrawable):
         self._create_buffered_strtree()
 
         return result
+
+    def __eq__(self, other):
+        if not isinstance(other, LaneletNetwork):
+            warnings.warn(f"Inequality between LaneletNetwork {repr(self)} and different type {type(other)}")
+            return False
+
+        list_elements_eq = True
+        lanelet_network_eq = True
+        elements = [self._lanelets, self._intersections, self._traffic_signs, self._traffic_lights]
+        elements_other = [other._lanelets, other._intersections, other._traffic_signs, other._traffic_lights]
+        for i in range(0, len(elements)):
+            e = elements[i]
+            e_other = elements_other[i]
+            lanelet_network_eq = lanelet_network_eq and len(e) == len(e_other)
+            for k in e.keys():
+                if k not in e_other:
+                    lanelet_network_eq = False
+                    continue
+                if e.get(k) != e_other.get(k):
+                    list_elements_eq = False
+
+        if not lanelet_network_eq:
+            warnings.warn(f"Inequality of LaneletNetwork {repr(self)} and the other one {repr(other)}")
+
+        return lanelet_network_eq and list_elements_eq
+
+    def __hash__(self):
+        return hash((frozenset(self._lanelets.items()), frozenset(self._intersections.items()),
+                     frozenset(self._traffic_signs.items()), frozenset(self._traffic_lights.items())))
+
+    def __str__(self):
+        return f"LaneletNetwork consists of lanelets {set(self._lanelets.keys())}, " \
+               f"intersections {set(self._intersections.keys())}, " \
+               f"traffic signs {set(self._traffic_signs.keys())}, and traffic lights {set(self._traffic_lights.keys())}"
+
+    def __repr__(self):
+        return f"LaneletNetwork(lanelets={repr(self._lanelets)}, intersections={repr(self._intersections)}, " \
+               f"traffic_signs={repr(self._traffic_signs)}, traffic_lights={repr(self._traffic_lights)})"
 
     @property
     def lanelets(self) -> List[Lanelet]:
@@ -1547,12 +1679,6 @@ class LaneletNetwork(IDrawable):
 
         # return sorted list
         return [lanelets[i] for i in indices]
-
-    def __str__(self):
-        return_str = ''
-        for lanelet_id in self._lanelets.keys():
-            return_str += '{:8d} lanelet\n'.format(lanelet_id)
-        return return_str
 
     def draw(self, renderer: IRenderer, draw_params: Union[ParamServer, dict, None] = None,
              call_stack: Optional[Tuple[str, ...]] = tuple()):