+++ commonroad/tests/scenario/test_intersection.py
@@ -64,6 +64,32 @@ class TestIntersection(unittest.TestCase):
         self.assertListEqual(list(exp_result.values()), [val.incoming_id for val
                                                          in self._intersection_1.map_incoming_lanelets.values()])
 
+    def test_equality(self):
+        incoming_1 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        incoming_2 = IntersectionIncomingElement(3, {20, 21}, {22, 23}, {24, 25}, {26, 27}, 28)
+        intersection_1 = Intersection(1, [incoming_1, incoming_2], {30, 31})
+        intersection_2 = Intersection(1, [incoming_1, incoming_2], {30, 31})
+        self.assertTrue(intersection_1 == intersection_2)
+
+        intersection_2 = Intersection(5, [incoming_1, incoming_2], {30, 31})
+        self.assertFalse(intersection_1 == intersection_2)
+
+        intersection_2 = Intersection(1, [incoming_1], {30, 31})
+        self.assertFalse(intersection_1 == intersection_2)
+
+        intersection_2 = Intersection(1, [incoming_1, incoming_2], {30, 31, 34})
+        self.assertFalse(intersection_1 == intersection_2)
+
+    def test_hash(self):
+        incoming = IntersectionIncomingElement(3, {20, 21}, {22, 23}, {24, 25}, {26, 27}, 28)
+        intersection_1 = Intersection(1, [incoming], {30, 31})
+        intersection_2 = Intersection(1, [incoming], {30, 31})
+        self.assertEqual(hash(intersection_1), hash(intersection_2))
+
+        incoming = IntersectionIncomingElement(3, {20, 21}, {22, 23}, {24, 25}, {26, 28}, 28)
+        intersection_2 = Intersection(1, [incoming], {30, 31})
+        self.assertNotEqual(hash(intersection_1), hash(intersection_2))
+
 
 class TestIntersectionIncomingElement(unittest.TestCase):
     def test_initialization_IntersectionIncomingElement(self):
@@ -89,6 +115,44 @@ class TestIntersectionIncomingElement(unittest.TestCase):
 
         incoming_id_2 = '3'
         self.assertRaises(AssertionError, IntersectionIncomingElement, incoming_id_2)
+
+    def test_equality(self):
+        incoming_1 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        self.assertTrue(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(1, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        self.assertFalse(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(2, {11, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        self.assertFalse(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 17}, {14, 15}, {16, 17}, 18)
+        self.assertFalse(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {5, 15}, {16, 17}, 18)
+        self.assertFalse(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {7, 17}, 18)
+        self.assertFalse(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 5)
+        self.assertFalse(incoming_1 == incoming_2)
+
+        incoming_1 = IntersectionIncomingElement(2, {10, 11}, None, {14, 15}, {16, 17}, 18)
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, None, {14, 15}, {16, 17}, 18)
+        self.assertTrue(incoming_1 == incoming_2)
+
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {1, 2, 3, 4}, {14, 15}, {16, 17}, 18)
+        self.assertFalse(incoming_1 == incoming_2)
+
+    def test_hash(self):
+        incoming_1 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        self.assertEqual(hash(incoming_1), hash(incoming_2))
+
+        incoming_2 = IntersectionIncomingElement(2, {10, 11}, {12, 17}, {14, 15}, {16, 17}, 18)
+        self.assertNotEqual(hash(incoming_1), hash(incoming_2))
 
 
 if __name__ == '__main__':