+++ commonroad/tests/scenario/test_laneletNetwork.py
@@ -350,10 +350,50 @@ class TestLaneletNetwork(unittest.TestCase):
         self.assertEqual(len(self.lanelet.traffic_signs), 0)
         self.assertEqual(len(self.lanelet.stop_line.traffic_sign_ref), 0)
 
-    def test_to_string(self):
-        self.assertMultiLineEqual(self.lanelet_network.__str__(),
-                                  '{:8d} lanelet\n'.format(5) +
-                                  '{:8d} lanelet\n'.format(6))
+    def test_equality_hash(self):
+        left_vertices = np.array([[0, 1], [1, 1], [2, 1]])
+        center_vertices = np.array([[0, 0], [1, 0], [2, 0]])
+        right_vertices = np.array([[0, -1], [1, -1], [2, -1]])
+        lanelet_id = 3
+        lanelet = Lanelet(left_vertices, center_vertices, right_vertices, lanelet_id)
+
+        incoming = IntersectionIncomingElement(2, {9, 11}, {12, 13}, {14, 15}, {16, 17}, 18)
+        intersection = Intersection(1, [incoming], {30, 31})
+
+        traffic_sign_max_speed = TrafficSignElement(TrafficSignIDGermany.MAX_SPEED, ["15"])
+        traffic_sign = TrafficSign(1, [traffic_sign_max_speed], {3}, np.array([10.0, 7.0]))
+
+        cycle = [TrafficLightCycleElement(TrafficLightState.GREEN, 2)]
+        traffic_light = TrafficLight(234, cycle, np.array([10., 10.]), 5)
+
+        lanelet_network_1 = LaneletNetwork()
+        lanelet_network_2 = LaneletNetwork()
+        for lanelet_network in [lanelet_network_1, lanelet_network_2]:
+            lanelet_network.add_lanelet(lanelet)
+            lanelet_network.add_intersection(intersection)
+            lanelet_network.add_traffic_sign(traffic_sign, {3})
+            lanelet_network.add_traffic_light(traffic_light, {3})
+        self.assertTrue(lanelet_network_1 == lanelet_network_2)
+        self.assertTrue(hash(lanelet_network_1), hash(lanelet_network_2))
+
+        lanelet_network_2.remove_lanelet(3)
+        self.assertFalse(lanelet_network_1 == lanelet_network_2)
+        self.assertNotEqual(hash(lanelet_network_1), hash(lanelet_network_2))
+
+        lanelet_network_2.add_lanelet(lanelet)
+        lanelet_network_2.add_intersection(Intersection(2, [incoming], {30, 31}))
+        self.assertFalse(lanelet_network_1 == lanelet_network_2)
+        self.assertNotEqual(hash(lanelet_network_1), hash(lanelet_network_2))
+
+        lanelet_network_2.remove_intersection(2)
+        lanelet_network_2.add_traffic_sign(TrafficSign(4, [traffic_sign_max_speed], {3}, np.array([10.0, 7.0])), {3})
+        self.assertFalse(lanelet_network_1 == lanelet_network_2)
+        self.assertNotEqual(hash(lanelet_network_1), hash(lanelet_network_2))
+
+        lanelet_network_2.remove_traffic_sign(4)
+        lanelet_network_2.remove_traffic_light(234)
+        self.assertFalse(lanelet_network_1 == lanelet_network_2)
+        self.assertNotEqual(hash(lanelet_network_1), hash(lanelet_network_2))
 
 
 if __name__ == '__main__':