+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [2021.4] - 2021-X-X
 ###Added
 - Polyline utility functions, e.g., resampling, path length, orientation, curvature, intersection
+- `__eq__` functions for LaneletNetwork and related classes (e.g., traffic sign, traffic light, stop line, etc.)
+
+### Fixed
+- Various small bug fixes
+
 
 ## [2021.3] - 2021-09-21
 ### Added